requirements_clean.txt
# Minimal requirements for SageVault RAG MVP
streamlit>=1.28.0
requests>=2.28.0
sentence-transformers>=2.2.0
chromadb==0.4.15
groq>=0.9.0
openai>=1.40.0
anthropic>=0.7.0
google-generativeai>=0.7.0
numpy>=1.24.0